Added render (); fixed backplate size
The back plate was sizing wrong; it had an extra wall width of depth. And needed to account for the pocket which isn't a wall thickness deep.

Added render() to all the final output in order to ensure that it's manifold. The STL processing was hit-and-miss in the slicers prior to that; now the final output is just the manifold exterior surfaces as expected.

This should now be perfectly functional as printed. It'll still require manual tweaking should you want to change parameters so be aware of that.